+++ src/assets/js/lumiere_classic_links.js
@@ -20,8 +20,10 @@ document.addEventListener(
 				var tmppopupLarg = lumiere_vars.popupLarg;
 				var tmppopupLong = lumiere_vars.popupLong;
 				// var mid from the class data-highslidepeople to build the link
-				var misc_term = jQuery( this ).closest( 'a' ).data( 'modal_window_people' );
-				var url_imdbperso = lumiere_vars.urlpopup_person + '?mid=' + misc_term;
+				var modal_window_people = jQuery( this ).closest( 'a' ).data( 'modal_window_people' );
+				var modal_window_nonce = jQuery( this ).closest( 'a' ).data( 'modal_window_nonce' );
+				
+				var url_imdbperso = lumiere_vars.urlpopup_person + '?mid=' + modal_window_people + '&_wpnonce=' + modal_window_nonce;
 
 				// classic popup
 				window.open( url_imdbperso, 'popup', 'resizable=yes, toolbar=no, scrollbars=yes, location=no, width=' + tmppopupLarg + ', height=' + tmppopupLong + ', top=5, left=5' )
@@ -36,8 +38,10 @@ document.addEventListener(
 				var tmppopupLarg = lumiere_vars.popupLarg;
 				var tmppopupLong = lumiere_vars.popupLong;
 				// var mid from the class data-highslidepeople to build the link
-				var misc_term = jQuery( this ).closest( 'a' ).data( 'modal_window_film' );
-				var url_imdbperso = lumiere_vars.urlpopup_film + '?film=' + misc_term;
+				var modal_window_film = jQuery( this ).closest( 'a' ).data( 'modal_window_film' );
+				var modal_window_nonce = jQuery( this ).closest( 'a' ).data( 'modal_window_nonce' );
+				
+				var url_imdbperso = lumiere_vars.urlpopup_film + '?film=' + modal_window_film + '&_wpnonce=' + modal_window_nonce;
 
 				// classic popup
 				window.open( url_imdbperso, 'popup', 'resizable=yes, toolbar=no, scrollbars=yes, location=no, width=' + tmppopupLarg + ', height=' + tmppopupLong + ', top=5, left=5' );
@@ -52,8 +56,10 @@ document.addEventListener(
 				var tmppopupLarg = lumiere_vars.popupLarg;
 				var tmppopupLong = lumiere_vars.popupLong;
 				// var mid from the class data-highslidepeople to build the link
-				var misc_term = jQuery( this ).closest( 'a' ).data( 'modal_window_filmid' );
-				var url_imdbperso = lumiere_vars.urlpopup_film + '?mid=' + misc_term;
+				var modal_window_filmid = jQuery( this ).closest( 'a' ).data( 'modal_window_filmid' );
+				var modal_window_nonce = jQuery( this ).closest( 'a' ).data( 'modal_window_nonce' );
+				
+				var url_imdbperso = lumiere_vars.urlpopup_film + '?mid=' + modal_window_filmid + '&_wpnonce=' + modal_window_nonce;
 
 				// classic popup
 				window.open( url_imdbperso, 'popup', 'resizable=yes, toolbar=no, scrollbars=yes, location=no, width=' + tmppopupLarg + ', height=' + tmppopupLong + ', top=5, left=5' );