+++ tests/configs_static_tools/constants.php
@@ -19,6 +19,14 @@ if ( ! defined( 'WP_PLUGIN_DIR' ) ) {
 	define( 'WP_PLUGIN_DIR', ABSPATH . 'wp-includes/plugins/lumiere-movies/');
 }
 
+if ( !defined( 'LUMIERE_WP_PATH' ) ) {
+	define( 'LUMIERE_WP_PATH', ABSPATH . 'wp-content/plugins/lumiere-movies/' );
+}
+
+if ( !defined( 'LUMIERE_WP_URL' ) ) {
+	define( 'LUMIERE_WP_URL', 'https://whatver.com/wp-content/plugins/lumiere-movies/' );
+}
+
 /**
  * The WP_DEBUG_LOG is either bool or string, but is defined as only bool in stubs
  * Using random function to have bool or string defines
@@ -34,11 +42,28 @@ if ( ! defined( 'WP_DEBUG_LOG' ) ) {
 	define( 'WP_DEBUG_LOG', $final );
 }
 
-/* Dynamic constant, declaration for Psalm but it brings a new error about always false in class logger...
+/* Dynamic constant, declaration for Psalm -> but then brings issue with PHPStan!
 if ( ! defined( 'DOING_CRON' ) ) {
 	define('DOING_CRON',
 	// @var mixed $x
 	$x = false
 	);
 }
-*/
+if ( ! defined( 'REST_REQUEST' ) ) {
+	define('REST_REQUEST',
+	// @var mixed $x
+	$x = false
+	);
+}
+if ( ! defined( 'DOING_AUTOSAVE' ) ) {
+	define('DOING_AUTOSAVE',
+	// @var mixed $x
+	$x = false
+	);
+}
+if ( ! defined( 'XMLRPC_REQUEST' ) ) {
+	define('XMLRPC_REQUEST',
+	// @var mixed $x
+	$x = false
+	);
+}*/